// src/models/amenity.ts
export interface IAmenity {
    id: number;
    lat: string;
    lng: string;
    name: string;
    housenumber: string;
    street: string;
    city: string;
    phone: string;
    email: string;
    country: string;
    website: string;
    openingHours: string;
    facebook: string;
    instagram: string;
    twitter: string;
    isFav: boolean;
}